refactor: Refactor to separate view and logic of text field
